// frontend/src/openapi/models/ConversationTurn.ts
export interface ConversationTurn {
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    traceId: string;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    conversationId: string;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    capability: string | null;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    status: string;
    /**
     *
     * @type {Date}
     * @memberof ConversationTurn
     */
    startedAt: Date | null;
    /**
     *
     * @type {number}
     * @memberof ConversationTurn
     */
    successScore: number | null;
    /**
     *
     * @type {number}
     * @memberof ConversationTurn
     */
    sessionScore: number | null;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    sessionRationale: string;
    /**
     *
     * @type {ExecutionIntent}
     * @memberof ConversationTurn
     */
    intent: ExecutionIntent | null;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    inputText: string;
    /**
     *
     * @type {string}
     * @memberof ConversationTurn
     */
    outputText: string;
    /**
     *
     * @type {ExecutionTaskState}
     * @memberof ConversationTurn
     */
    taskState: ExecutionTaskState | null;
    /**
     *
     * @type {any}
     * @memberof ConversationTurn
     */
    stepResults: any | null;
    /**
     *
     * @type {Array<ExecutionTool>}
     * @memberof ConversationTurn
     */
    tools: Array<ExecutionTool>;
}